+++ app/src/main/java/org/breezyweather/remoteviews/presenters/ClockDayWeekWidgetIMP.kt
@@ -20,6 +20,7 @@ import android.appwidget.AppWidgetManager
 import android.content.ComponentName
 import android.content.Context
 import android.graphics.Color
+import android.os.Build
 import android.util.TypedValue
 import android.view.View
 import android.widget.RemoteViews
@@ -39,6 +40,7 @@ import org.breezyweather.remoteviews.Widgets
 import org.breezyweather.ui.theme.resource.ResourceHelper
 import org.breezyweather.ui.theme.resource.ResourcesProviderFactory
 import java.util.Date
+import kotlin.math.roundToInt
 
 object ClockDayWeekWidgetIMP : AbstractRemoteViewsPresenter() {
 
@@ -294,6 +296,13 @@ object ClockDayWeekWidgetIMP : AbstractRemoteViewsPresenter() {
                 setTextViewTextSize(R.id.widget_clock_day_week_temp_3, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_clock_day_week_temp_4, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_clock_day_week_temp_5, TypedValue.COMPLEX_UNIT_PX, contentSize)
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                    setInt(R.id.widget_clock_day_week_temp_1, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_clock_day_week_temp_2, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_clock_day_week_temp_3, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_clock_day_week_temp_4, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_clock_day_week_temp_5, "setLineHeight", contentSize.roundToInt())
+                }
             }
         }
         if (color.showCard) {