+++ app/src/main/java/org/breezyweather/remoteviews/presenters/DayWeekWidgetIMP.kt
@@ -20,6 +20,7 @@ import android.appwidget.AppWidgetManager
 import android.content.ComponentName
 import android.content.Context
 import android.graphics.Color
+import android.os.Build
 import android.util.TypedValue
 import android.view.View
 import android.widget.RemoteViews
@@ -49,6 +50,7 @@ import org.breezyweather.ui.theme.resource.ResourceHelper
 import org.breezyweather.ui.theme.resource.ResourcesProviderFactory
 import org.breezyweather.ui.theme.resource.providers.ResourceProvider
 import java.util.Date
+import kotlin.math.roundToInt
 
 object DayWeekWidgetIMP : AbstractRemoteViewsPresenter() {
 
@@ -192,6 +194,13 @@ object DayWeekWidgetIMP : AbstractRemoteViewsPresenter() {
                 setTextViewTextSize(R.id.widget_day_week_temp_3, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_day_week_temp_4, TypedValue.COMPLEX_UNIT_PX, contentSize)
                 setTextViewTextSize(R.id.widget_day_week_temp_5, TypedValue.COMPLEX_UNIT_PX, contentSize)
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                    setInt(R.id.widget_day_week_temp_1, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_day_week_temp_2, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_day_week_temp_3, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_day_week_temp_4, "setLineHeight", contentSize.roundToInt())
+                    setInt(R.id.widget_day_week_temp_5, "setLineHeight", contentSize.roundToInt())
+                }
             }
         }