+++ app/src/main/java/org/breezyweather/remoteviews/presenters/DayWidgetIMP.kt
@@ -20,6 +20,7 @@ import android.appwidget.AppWidgetManager
 import android.content.ComponentName
 import android.content.Context
 import android.graphics.Color
+import android.os.Build
 import android.util.TypedValue
 import android.view.View
 import android.widget.RemoteViews
@@ -296,6 +297,18 @@ object DayWidgetIMP : AbstractRemoteViewsPresenter() {
                     TypedValue.COMPLEX_UNIT_PX,
                     getTimeSize(context, viewStyle).times(textSize).div(100f)
                 )
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                    setInt(
+                        R.id.widget_day_title,
+                        "setLineHeight",
+                        getTitleSize(context, viewStyle).times(textSize).div(100f).roundToInt()
+                    )
+                    setInt(
+                        R.id.widget_day_subtitle,
+                        "setLineHeight",
+                        getSubtitleSize(context, viewStyle).times(textSize).div(100f).roundToInt()
+                    )
+                }
             }
         }
         views.setViewVisibility(R.id.widget_day_time, if (hideSubtitle) View.GONE else View.VISIBLE)